fix(): changed int to long and minor changes

+++ app/src/main/java/com/example/census/model/CensusForm.java
@@ -2,9 +2,9 @@ package com.example.census.model;
 
 public class CensusForm {
     private int id;
-    private int citizen_tin;
+    private long citizen_tin;
     private int age;
-    private int number;
+    private long number;
     private String sex;
     private String birth;
     private String owner_rel;
@@ -62,11 +62,11 @@ public class CensusForm {
         this.id = id;
     }
 
-    public int getCitizen_tin() {
+    public long getCitizen_tin() {
         return citizen_tin;
     }
 
-    public void setCitizen_tin(int citizen_tin) {
+    public void setCitizen_tin(long citizen_tin) {
         this.citizen_tin = citizen_tin;
     }
 
@@ -78,11 +78,11 @@ public class CensusForm {
         this.age = age;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
